+++ src/main/java/turniplabs/halplibe/util/OptionsInitEntrypoint.java
@@ -6,5 +6,5 @@ public interface OptionsInitEntrypoint {
     /**
      * The entrypoint name inside {@code fabric.mod.json} is {@code initOptions}.
      */
-    void initOptions(GameSettings settings);
+    void initOptions();
 }